% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bdf_native.R
\name{digital_to_physical}
\alias{digital_to_physical}
\title{Convert Digital Values to Physical Units}
\usage{
digital_to_physical(
  digital,
  digital_min,
  digital_max,
  physical_min,
  physical_max
)
}
\arguments{
\item{digital}{Integer vector of digital values}

\item{digital_min}{Digital minimum from header}

\item{digital_max}{Digital maximum from header}

\item{physical_min}{Physical minimum from header (microV)}

\item{physical_max}{Physical maximum from header (microV)}
}
\value{
Numeric vector of physical values
}
\description{
Convert Digital Values to Physical Units
}
\keyword{internal}
