% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_info.R
\name{get_biosemi_channel_info}
\alias{get_biosemi_channel_info}
\title{Get BioSemi Channel Information}
\usage{
get_biosemi_channel_info(n_channels)
}
\arguments{
\item{n_channels}{Number of channels in the recording}
}
\value{
Data frame with columns: channel_number, name, position_type
}
\description{
Returns information about standard BioSemi channel layouts
(64-channel, 128-channel, 256-channel systems)
}
\details{
BioSemi systems typically use standard 10-20/10-10 electrode positions.
This function provides information about standard configurations.
}
\examples{
\dontrun{
  biosemi_64_info <- get_biosemi_channel_info(64)
}

}
