% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bdf_events.R
\name{plot_bdf_events}
\alias{plot_bdf_events}
\title{Plot BDF event timeline}
\usage{
plot_bdf_events(events, show_codes = FALSE)
}
\arguments{
\item{events}{Data frame returned from \code{extract_bdf_events()}}

\item{show_codes}{Logical. If TRUE, label each event with its trigger code.}
}
\description{
Plot BDF event timeline
}
