% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bdf_events.R
\name{summary_bdf_events}
\alias{summary_bdf_events}
\title{Summary of extracted BDF events}
\usage{
summary_bdf_events(events)
}
\arguments{
\item{events}{Data frame returned from \code{extract_bdf_events()}}
}
\description{
Summary of extracted BDF events
}
