% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{unpack_int24}
\alias{unpack_int24}
\title{Internal: unpack 24-bit signed integers from raw vector}
\usage{
unpack_int24(raw_vec)
}
\description{
Internal: unpack 24-bit signed integers from raw vector
}
\keyword{internal}
