% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bdf_native.R
\name{unpack_int24_native}
\alias{unpack_int24_native}
\title{Unpack 24-bit Signed Integers}
\usage{
unpack_int24_native(raw_vec)
}
\arguments{
\item{raw_vec}{Raw vector of bytes (length must be multiple of 3)}
}
\value{
Integer vector of unpacked values
}
\description{
Unpack 24-bit Signed Integers
}
\keyword{internal}
