% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{.eog_resolve_picks}
\alias{.eog_resolve_picks}
\title{Turn Channel Picks Into Channel Names (internal)}
\usage{
.eog_resolve_picks(picks, eeg, arg)
}
\arguments{
\item{picks}{Character vector of channel names, or numeric vector of
channel indices.}

\item{eeg}{An object of class 'eeg' or 'eeg_epochs'.}

\item{arg}{Name of the calling argument, used in error messages.}
}
\value{
Character vector of channel names, in the order given.
}
\description{
Turn Channel Picks Into Channel Names (internal)
}
\keyword{internal}
